renamed racoon package to statement package
