+++ src/main/kotlin/it/zoo/vendro/racoon/habitat/statements/parameters/ParameterMapping.kt
@@ -1,4 +1,4 @@
-package it.zoo.vendro.racoon.habitat.racoons.parameters
+package it.zoo.vendro.racoon.habitat.statements.parameters
 
 import java.sql.SQLException
